+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New Feature
 
 - Added component `DiagramNode` for CA Learning graph
+- Added styling for `DiagramNode` when selected in CA Learning graph
 
 ### Bug Fix